内存受限允许杀死keepalive应用

+++ services/abilitymgr/src/ability_manager_service.cpp
@@ -5703,7 +5703,7 @@ int AbilityManagerService::KillProcess(const std::string &bundleName)
         return GET_BUNDLE_INFO_FAILED;
     }
 
-    if (bundleInfo.isKeepAlive) {
+    if (bundleInfo.isKeepAlive && DelayedSingleton<AppScheduler>::GetInstance()->IsMemorySizeSufficent()) {
         TAG_LOGE(AAFwkTag::ABILITYMGR, "Can not kill keep alive process.");
         return KILL_PROCESS_KEEP_ALIVE;
     }
@@ -9691,6 +9691,14 @@ void AbilityManagerService::NotifyConfigurationChange(const AppExecFwk::Configur
         return;
     }
     collaborator->UpdateConfiguration(config, userId);
+}
+
+void AbilityManagerService::NotifyStartResidentProcess(std::vector<AppExecFwk::BundleInfo> &bundleInfos)
+{
+    DelayedSingleton<ResidentProcessManager>::GetInstance()->StartResidentProcessWithMainElement(bundleInfos);
+    if (!bundleInfos.empty()) {
+        DelayedSingleton<ResidentProcessManager>::GetInstance()->StartResidentProcess(bundleInfos);
+    }
 }
 
 int32_t AbilityManagerService::OpenFile(const Uri& uri, uint32_t flag)
